Lhvt (#2)
* Merged the two patch branches into master

* Changed the default compiler to gcc. Added optimization flags

* This is the current stage of public SAGE (assuming the two PRs do not change drastically). Plus, a check to stop merger times from unphysical.

* Added a flyby flag into SAGE. Will be 0 for LHalotree based codes and 1 for trees that can handle multiple redshift zero FOF roots. My treeconverter code will get updated to output this extra bit of info within MostBoundID

* Fixed the data-types in the associated python routines

* Added the -Wpadded warning back in. Removed one plot from allresults

* Commented out the -Wpadded. Struct definition is mostly sorted for now. Increased number of memory blocks to handle Vishnu full-res

* Pulling in recent updates (#2)

* Reallocating Fof_Maxgals such that high resolution sims can be accommodated easily (instead of crashing). Fixes Bolshoi.

* Notebook updated to work with latest version of model and better match figures in paper

* One extra instruction for the notebook

* Updated the mini-MR treefile download link in README.md

* Plot/read scripts for iPython Notebook added

* Started work on HDF5 reader.

Need to check that reading of an array of attributes is working fine.
Since reading an array happens multiple times (for the data as well) maybe
should put in its own function.

* Reading in of table is correct.

Want to generalise so the user can specify the name of the fields of the
HDF5 file.

* Reworked directory structure.

Moved files dealing with io to their own directory with their own header
files.  In the final verison, core_io_tree.c should only act as a
wrapper file, calling the appropriate binary or HDF5 versions within io/
directory.

* Made io files more consistent with respective to local/external tags.

* Function that populates the metadata names works properly.

Have also added the -Werror option to the Makefile so I don't miss any
errors.

* Starting working on the reading of the actual tree.

* Hdf5 writer (#1)

* Added Simon hdf5 file writer

* Reading in of table is correct.

Want to generalise so the user can specify the name of the fields of the
HDF5 file.

* Committing all the changes. Need to add in the changes to core_io_tree and make those changes in sync with Jacobs reader section

* Fixing up compile errors

* Added object files and emacs backup files to gitigore

* Compiles and runs on g2 after merge.

* HDF5 reader now works fully.

The read in of the tree is done in an incredibly sloppy manner.  This
was done mainly because I wanted to get it done and ensured it works.
In future, this should be updated to include macros/other functions so
code doesn't need to be copy pasted.

The code has been checked by running with binary mode and HDF5 mode and
doing a diff of the outputs.  This resulted in no differences.  I'm
about to check back out master, run the base version and do a final
check to ensure nothing is different there.

* Reverted the accidental changes to allresults.

* make clean removes the .exe. Claned up parameter file.

* make clean removes the .exe. Cleaned up parameter file.

* Cleaned up in response to Manodeep's comments.

Need to give a check over the entire code for validation of fwrites and
mallocs.  Actually I think the validation should just be put into
myfwrite and mymalloc and ensure that everything uses those functions.

* HDF5 reader now uses Macros.

Instead of copy pasting the code like a noob I've replaced it all with
two macros. I needed 2 because of fields such as Pos which has multiple
dimesions.

* Models are now properly bitwise identical.

I have yet to implement the comments that Manodeep made ~1 hr ago.

* Added default for switch cases.

* Initialized output_galaxy.

Comparison to Master (after initializing output_galaxy as well) is now
bitwise identical for both binary and HDF5 readers.  Valgrind produces
no memory leaks, warnings or errors.

* Initialize galaxy_output.

Since struct(GALAXY_OUTPUT) is not aligned to 8 bytes it contains a pad.
When writing out the galaxies uninitialized bytes are then written out.
This commit fixes this by initializing the struct (including the pad) to
0.

Also properly free the GSL random_generator.

Valgrind produces no memory leaks, warnings or errors.